Make test env consistent with dev env
Comment out the test environment config as it is causing problems
with not behaving like the development app. The test environment
is setup for using Ember's own testing, but this doesn't work
well with the fact that database-stored test data and transactions
are easily managed within the Rails app. For now, proceed with all
tests being performed by the Rails app. Try to get Ember tests
in where appropriate.